packages/brain/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fme-brain"
version = "0.1.0"
description = "Local LLM pipeline: sales chat -> seller next-message suggestion (JSON)."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "openai>=1.0.0",
  "pydantic>=2.0",
  "python-dotenv>=1.0.0",
  "typer>=0.12.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
fme-brain = "fme_brain.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/fme_brain"]

[tool.pytest.ini_options]
testpaths = ["tests"]
